# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "WaveOne API",
        "contact": {
            "name": "Aneep Tandel",
            "url": "https://github.com/aneepct"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/farm_activity/all.{format?}": {
            "get": {
                "tags": [
                    "Farm Activities"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns all farm activities.",
                "operationId": "api.farm_activity.index",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Settings."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/farm_activity/{id}/get_activity.{format?}": {
            "get": {
                "tags": [
                    "Farm Activities"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns all farm activities.",
                "operationId": "api.farm_activity.show",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Farm Activity Id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Settings."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/farm_activity/create.{format?}": {
            "post": {
                "tags": [
                    "Farm Activities"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Removefarm activity.",
                "operationId": "api.farm_activity.store",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Activity Name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Activity Code.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Settings."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/farm_activity/{id}/update_farm_acitvity.{format?}": {
            "put": {
                "tags": [
                    "Farm Activities"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Update Farm activities.",
                "operationId": "api.farm_activity.update",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of activity",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of activity",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Code of activity",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Settings."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/farm_activity/{id}/delete.{format?}": {
            "delete": {
                "tags": [
                    "Farm Activities"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Delete Farm activity.",
                "operationId": "api.farm_activity.destroy",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of activity",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App Settings."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        }
    },
...
}